+++ pkg/idp/idp.go
@@ -71,6 +71,13 @@ func New(c Config) (*Service, error) {
 }
 
 func (s *Service) Reconcile(ctx context.Context) error {
+	// We do not handle apps that have user configmaps set up due to a bug where configuration in secrets can be overwritten
+	//TODO: solve this gracefully
+	if userConfigMapPresent(s.app) {
+		s.log.Info("Dex app has a user configmap set up for configuration. Cancelling reconcillation. We recommend to move configuration to a user secret.")
+		return s.ReconcileDelete(ctx)
+	}
+
 	// Add secret config to app instance
 	dexSecretConfig := GetDexSecretConfig(s.app.Namespace)
 	if !dexSecretConfigIsPresent(s.app, dexSecretConfig) {